코드 정렬

+++ products/views.py
@@ -55,8 +55,9 @@ class ProductListAPIView(APIView):
     def post(self, request):
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
-            serializer.save(author=request.user)  # 작성자 필드에 로그인된 사용자 할당
+            serializer.save(author=request.user)  
             return Response(serializer.data, status=201) 
+
 
 
 
@@ -91,7 +92,7 @@ class ProductListAPIView(APIView):
 
 #클래스형 상세조회, 수정, 삭제
 class ProductDetailAPIView(APIView):
-    permission_classes = [IsAuthenticated]  # 클래스 속성으로 이동
+    permission_classes = [IsAuthenticated]  
 
     def get_object(self, pk):
         return get_object_or_404(Product, pk=pk)
@@ -145,19 +146,22 @@ class ProductDetailAPIView(APIView):
 
 #클래스형  코멘트 조회, 생성  
 class CommentListAPIView(APIView):
+    
+
     def get(self, request, pk):
         product = get_object_or_404(Product, pk=pk)
         comments = product.comments.all()
         serializer = CommentSerializer(comments, many=True)
         return Response(serializer.data)
     
-    permission_classes=[IsAuthenticated]
+    permission_classes = [IsAuthenticated]  
     def post(self, request, pk):
         serializer = CommentSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors)
+
 
 
 #함수형 코멘트 삭제, 수정
@@ -181,21 +185,42 @@ class CommentListAPIView(APIView):
 
 #클래스형 삭제 수정
 class CommentDetailAPIView(APIView):
-    permission_classes=[IsAuthenticated]
-    
+    permission_classes = [IsAuthenticated]
+
     def get_object(self, pk):
         return get_object_or_404(Comment, pk=pk)
 
     def delete(self, request, pk):
         comment = self.get_object(pk)
+        # 댓글 작성자가 현재 사용자와 같은지 확인
+        if request.user != comment.author:
+            raise PermissionDenied("You do not have permission to delete this comment.")
         comment.delete()
-        data={"delete":f"Comment({pk}) is deleted."}
+        data = {"delete": f"Comment({pk}) is deleted."}
         return Response(data, status=status.HTTP_200_OK)
-    
+
     def put(self, request, pk):
         comment = self.get_object(pk)
-        serializer=CommentSerializer(comment, data=request.data, partial=True)
+        # 댓글 작성자가 현재 사용자와 같은지 확인
+        if request.user != comment.author:
+            raise PermissionDenied("You do not have permission to edit this comment.")
+        serializer = CommentSerializer(comment, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
